elixir: Day5 solutions.
day5_bad is the first attempt which failed miserably.
Basically I was trying to find intersecting line segments instead of
marking all individual points on the line.
For the life of me, I can't figure out why the code give 6440 as the answer.
Maybe someday I will come back to this and fix it...